remove homefragment
