+++ Singers/app/src/main/java/ru/daria/singers/MainFragment.java
@@ -2,18 +2,29 @@ package ru.daria.singers;
 
 import android.app.Activity;
 import android.app.Fragment;
+import android.app.FragmentManager;
+import android.app.FragmentTransaction;
+import android.app.LoaderManager;
+import android.content.Loader;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.LinearLayout;
 import android.widget.ListView;
+import android.widget.ProgressBar;
+import android.widget.TextView;
 
 import com.nhaarman.listviewanimations.appearance.simple.SwingRightInAnimationAdapter;
 import com.nostra13.universalimageloader.core.ImageLoader;
 import com.nostra13.universalimageloader.core.ImageLoaderConfiguration;
 
+import java.io.IOException;
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -22,13 +33,67 @@ import butterknife.ButterKnife;
 /**
  * Created by dsukmanova on 15.07.16.
  */
-public class MainFragment extends Fragment implements View.OnClickListener {
+public class MainFragment extends Fragment implements View.OnClickListener, LoaderManager.LoaderCallbacks<DataLoaderResult<List<Singer>>> {
     List<Singer> singers = new ArrayList<>();
     ListItemAdapter adapter;
     OnSingerSelectedListener mListener;
+    final int LOADER_ID = 0;
 
     public interface OnSingerSelectedListener {
         void onSingerSelected(Singer singer);
+    }
+
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        Loader loader = getLoaderManager().initLoader(LOADER_ID, null, this);
+        loader.forceLoad();
+    }
+
+    @Override
+    public Loader<DataLoaderResult<List<Singer>>> onCreateLoader(int id, Bundle args) {
+        DataLoader loader = new DataLoader(getActivity());
+        return loader;
+    }
+
+    @Override
+    public void onLoadFinished(Loader<DataLoaderResult<List<Singer>>> loader, DataLoaderResult<List<Singer>> data) {
+        Exception ex = data.getException();
+        if (ex != null) {
+            ex.printStackTrace();
+            if (ex instanceof IOException) {
+                /*Button b = ButterKnife.findById(view, R.id.tryAgainButton);
+                b.setVisibility(View.VISIBLE);
+                b.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        Button b = (Button) v;
+                        b.setVisibility(View.INVISIBLE);
+                        getLoaderManager().restartLoader(LOADER_ID, null, HomeFragment.this);
+                    }
+                });
+                TextView v = ButterKnife.findById(view, R.id.splashText);
+                v.setText("Невозможно загрузить список исполнителей.");*/
+                ex.printStackTrace();
+            }
+        } else {
+
+            singers = data.getResult();
+            MainFragment fragment = new MainFragment();
+            Bundle arg = new Bundle();
+            arg.putSerializable("SINGERS", (Serializable) singers);
+            fragment.setArguments(arg);
+
+            FragmentManager fragmentManager = getFragmentManager();
+            FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+            fragmentTransaction.replace(R.id.fragmentContainer, fragment);
+            fragmentTransaction.commitAllowingStateLoss();
+        }
+    }
+
+    @Override
+    public void onLoaderReset(Loader<DataLoaderResult<List<Singer>>> loader) {
+
     }
 
     @Override
